+++ app/components/children/AnnouncementsView.js
@@ -1,38 +1,6 @@
 var React = require("react");
-var helpers = require("../utils/helpers");
 
 var AnnouncementsView = React.createClass({
-    getInitialState: function() {
-        return {
-            title: "",
-            content: "",
-            datetime: ""
-        };
-    },
-
-    componentDidMount: function() {
-        this.getAnnouncements();
-    },
-
-    // componentDidUpdate: function () {
-    //     this.getAnnouncements();
-    // },
-
-    getAnnouncements: function() {
-        helpers.getAnnouncements().then(function(response) {
-        //   console.log(response.data[response.data.length - 1]);
-          this.setState({
-            title: response.data[response.data.length -1].title,
-            content: response.data[response.data.length -1].content
-          });
-
-        //   if (response !== this.state.allEmployees) {
-        //       this.setState({ allEmployees: response.data });
-        //       this.activeButtons();
-        //   }
-        }.bind(this));
-    },
-
     render: function() {
         return (
             <div className="card-panel">
@@ -43,8 +11,10 @@ var AnnouncementsView = React.createClass({
                 </div>
                 <div className="row">
                     <div className="col s12">
-                        <h5>{this.state.title}</h5>
-                        <p>{this.state.content}</p>
+                        {/* <h5>{this.state.title}</h5> */}
+                        <h5>{this.props.title}</h5>
+                        <p>{this.props.content}</p>
+                        {/* <p>{this.state.content}</p> */}
                     </div>
                 </div>
             </div>
